requirements.txt
google-generativeai
chromadb
matplotlib
seaborn
langchain_google_genai
langchain-chroma
langchain_community
pypdf
FAISS
langchain_text_splitters